split out functions into separate files for easier testing
